+++ public/index.html
@@ -97,7 +97,7 @@
     </div>
   </div>
 
-  <script src="js/client.js" type="module"></script>
+  <script src="js/main.js" type="module"></script>
 
 </body>
